Move PopupScreen base class into PopupScreens.cpp/h
