+++ Common/UI/UIScreen.cpp
@@ -6,12 +6,10 @@
 #include "Common/System/Request.h"
 #include "Common/Input/InputState.h"
 #include "Common/Input/KeyCodes.h"
-#include "Common/Math/curves.h"
 #include "Common/UI/UIScreen.h"
 #include "Common/UI/Context.h"
 #include "Common/UI/Screen.h"
 #include "Common/UI/Root.h"
-#include "Common/Data/Text/I18n.h"
 #include "Common/Render/DrawBuffer.h"
 
 static const bool ClickDebug = false;
@@ -299,170 +297,4 @@ void UIScreen::OnOK(UI::EventParams &e) {
 
 void UIScreen::OnCancel(UI::EventParams &e) {
 	TriggerFinish(DR_CANCEL);
-}
-
-PopupScreen::PopupScreen(std::string_view title, std::string_view button1, std::string_view button2)
-	: title_(title) {
-	auto di = GetI18NCategory(I18NCat::DIALOG);
-	if (!button1.empty())
-		button1_ = di->T(button1);
-	if (!button2.empty())
-		button2_ = di->T(button2);
-	alpha_ = 0.0f;  // inherited
-}
-
-void PopupScreen::touch(const TouchInput &touch) {
-	if (!box_ || (touch.flags & TOUCH_DOWN) == 0) {
-		// Handle down-presses here.
-		UIDialogScreen::touch(touch);
-		return;
-	}
-
-	// Extra bounds to avoid closing the dialog while trying to aim for something
-	// near the edge. Now that we only close on actual down-events, we can shrink
-	// this border a bit.
-	if (!box_->GetBounds().Expand(30.0f, 30.0f).Contains(touch.x, touch.y)) {
-		TriggerFinish(DR_CANCEL);
-	}
-
-	UIDialogScreen::touch(touch);
-}
-
-bool PopupScreen::key(const KeyInput &key) {
-	if (key.flags & KEY_DOWN) {
-		if ((key.keyCode == NKCODE_ENTER || key.keyCode == NKCODE_NUMPAD_ENTER) && defaultButton_) {
-			UI::EventParams e{};
-			defaultButton_->OnClick.Trigger(e);
-			return true;
-		}
-	}
-
-	return UIDialogScreen::key(key);
-}
-
-void PopupScreen::update() {
-	UIDialogScreen::update();
-
-	if (defaultButton_)
-		defaultButton_->SetEnabled(CanComplete(DR_OK));
-
-	float animatePos = 1.0f;
-
-	++frames_;
-	if (finishFrame_ >= 0) {
-		float leadOut = bezierEaseInOut((frames_ - finishFrame_) * (1.0f / (float)FRAMES_LEAD_OUT));
-		animatePos = 1.0f - leadOut;
-
-		if (frames_ >= finishFrame_ + FRAMES_LEAD_OUT) {
-			// Actual finish happens here.
-			screenManager()->finishDialog(this, finishResult_);
-		}
-	} else if (frames_ < FRAMES_LEAD_IN) {
-		float leadIn = bezierEaseInOut(frames_ * (1.0f / (float)FRAMES_LEAD_IN));
-		animatePos = leadIn;
-	}
-
-	if (animatePos < 1.0f) {
-		alpha_ = animatePos;
-		scale_.x = 0.9f + animatePos * 0.1f;
-		scale_.y =  0.9f + animatePos * 0.1f;
-
-		if (hasPopupOrigin_) {
-			float xoff = popupOrigin_.x - g_display.dp_xres / 2;
-			float yoff = popupOrigin_.y - g_display.dp_yres / 2;
-
-			// Pull toward the origin a bit.
-			translation_.x = xoff * (1.0f - animatePos) * 0.2f;
-			translation_.y = yoff * (1.0f - animatePos) * 0.2f;
-		} else {
-			translation_.y = -g_display.dp_yres * (1.0f - animatePos) * 0.2f;
-		}
-	} else {
-		alpha_ = 1.0f;
-		scale_.x = 1.0f;
-		scale_.y = 1.0f;
-		translation_.x = 0.0f;
-		translation_.y = 0.0f;
-	}
-}
-
-void PopupScreen::SetPopupOrigin(const UI::View *view) {
-	hasPopupOrigin_ = true;
-	popupOrigin_ = view->GetBounds().Center();
-}
-
-void PopupScreen::TriggerFinish(DialogResult result) {
-	if (CanComplete(result)) {
-		ignoreInput_ = true;
-		finishFrame_ = frames_;
-		finishResult_ = result;
-
-		OnCompleted(result);
-	}
-	// Inform UI that popup close to hide OSK (if visible)
-	System_NotifyUIEvent(UIEventNotification::POPUP_CLOSED);
-}
-
-void PopupScreen::CreateViews() {
-	using namespace UI;
-	UIContext &dc = *screenManager()->getUIContext();
-
-	AnchorLayout *anchor = new AnchorLayout(new LayoutParams(FILL_PARENT, FILL_PARENT));
-	anchor->Overflow(false);
-	root_ = anchor;
-
-	float yres = screenManager()->getUIContext()->GetBounds().h;
-
-	AnchorLayoutParams *anchorParams;
-	if (!alignTop_) {
-		// Standard centering etc.
-		anchorParams = new AnchorLayoutParams(PopupWidth(), FillVertical() ? yres - 30 : WRAP_CONTENT,
-			dc.GetBounds().centerX(), dc.GetBounds().centerY() + offsetY_, NONE, NONE, true);
-	} else {
-		// Top-aligned, for dialogs where we need to pop a keyboard below.
-		anchorParams = new AnchorLayoutParams(PopupWidth(), FillVertical() ? yres - 30 : WRAP_CONTENT,
-			NONE, 0, NONE, NONE, false);
-	}
-
-	box_ = new LinearLayout(ORIENT_VERTICAL, anchorParams);
-
-	root_->Add(box_);
-	box_->SetBG(dc.GetTheme().popupStyle.background);
-	box_->SetHasDropShadow(hasDropShadow_);
-	// Since we scale a bit, make the dropshadow bleed past the edges.
-	box_->SetDropShadowExpand(std::max(g_display.dp_xres, g_display.dp_yres));
-	box_->SetSpacing(0.0f);
-
-	if (HasTitleBar()) {
-		View *title = new PopupHeader(title_);
-		box_->Add(title);
-	}
-
-	CreatePopupContents(box_);
-	root_->Recurse([](View *view) {
-		view->SetPopupStyle(true);
-	});
-
-	root_->SetDefaultFocusView(box_);
-	if (ShowButtons() && !button1_.empty()) {
-		// And the two buttons at the bottom.
-		LinearLayout *buttonRow = new LinearLayout(ORIENT_HORIZONTAL, new LinearLayoutParams(200, WRAP_CONTENT));
-		buttonRow->SetSpacing(0);
-		Margins buttonMargins(5, 5);
-
-		// Adjust button order to the platform default.
-		if (System_GetPropertyBool(SYSPROP_OK_BUTTON_LEFT)) {
-			defaultButton_ = buttonRow->Add(new Button(button1_, new LinearLayoutParams(1.0f, buttonMargins)));
-			defaultButton_->OnClick.Handle<UIScreen>(this, &UIScreen::OnOK);
-			if (!button2_.empty())
-				buttonRow->Add(new Button(button2_, new LinearLayoutParams(1.0f, buttonMargins)))->OnClick.Handle<UIScreen>(this, &UIScreen::OnCancel);
-		} else {
-			if (!button2_.empty())
-				buttonRow->Add(new Button(button2_, new LinearLayoutParams(1.0f)))->OnClick.Handle<UIScreen>(this, &UIScreen::OnCancel);
-			defaultButton_ = buttonRow->Add(new Button(button1_, new LinearLayoutParams(1.0f)));
-			defaultButton_->OnClick.Handle<UIScreen>(this, &UIScreen::OnOK);
-		}
-
-		box_->Add(buttonRow);
-	}
 }
