+++ UI/ControlMappingScreen.h
@@ -26,6 +26,7 @@
 
 #include "Common/UI/View.h"
 #include "Common/UI/UIScreen.h"
+#include "Common/UI/PopupScreens.h"
 #include "Common/Data/Text/I18n.h"
 
 #include "Core/ControlMapper.h"
@@ -64,7 +65,7 @@ private:
 	bool categoryToggles_[10]{};
 };
 
-class KeyMappingNewKeyDialog : public PopupScreen {
+class KeyMappingNewKeyDialog : public UI::PopupScreen {
 public:
 	explicit KeyMappingNewKeyDialog(int btn, bool replace, std::function<void(KeyMap::MultiInputMapping)> callback, I18NCat i18n)
 		: PopupScreen(T(i18n, "Map Key"), "Cancel", ""), pspBtn_(btn), callback_(callback) {}
@@ -97,7 +98,7 @@ private:
 	double delayUntil_ = 0.0f;
 };
 
-class KeyMappingNewMouseKeyDialog : public PopupScreen {
+class KeyMappingNewMouseKeyDialog : public UI::PopupScreen {
 public:
 	KeyMappingNewMouseKeyDialog(int btn, bool replace, std::function<void(KeyMap::MultiInputMapping)> callback, I18NCat i18n)
 		: PopupScreen(T(i18n, "Map Mouse"), "", ""), callback_(callback) {}
